util: extract LegacyFramingParser from Messages to isolate transitional code
Fixes #3465

Pure code-motion. Moves the legacy <message type=X>...</message> framing
parser introduced in #3463 (parseLegacyFraming, stripLegacyFraming,
cleanFramedContent, FramedSegment, three Pattern constants, MAX_FRAMING_DEPTH)
out of Messages.java into a dedicated ai.brokk.util.LegacyFramingParser class.

Why: Messages.java already mixes LLM-side rendering (format, getRepr) with
display-side rendering (formatForDisplay, getReprForDisplay). The legacy
parser is a third concern -- inverse parsing of a deprecated persisted
format -- with a planned removal point (#3462). Co-locating it with the
permanent rendering API made the class harder to navigate (~371 lines,
22 public methods) and obscured which symbols are transitional vs core.

Public API on the new class:

- LegacyFramingParser.parse(String) -- was Messages.parseLegacyFraming
- LegacyFramingParser.strip(String) -- was Messages.stripLegacyFraming
- LegacyFramingParser.FramedSegment -- was Messages.FramedSegment

Class-level javadoc documents the transitional nature with a TODO(#3462)
removal note, so a future maintainer cleaning up after structured-message
persistence lands has an explicit deletion target.

Call sites updated:

- TaskEntry.mopMessages
- MOPBridge.sendHistoryTask
- BrokkAcpAgent.scheduleConversationReplay

Test class renamed MessagesLegacyFramingTest -> LegacyFramingParserTest;
ArrayDeque + Pattern imports removed from Messages.java now that they're
unused.

Behavior is unchanged -- all 13 existing parser tests pass on the new
class without modification.
