+++ app/src/main/java/ai/brokk/util/Messages.java
@@ -9,13 +9,11 @@ import ai.brokk.context.ContextFragment;
 import dev.langchain4j.agent.tool.ToolExecutionRequest;
 import dev.langchain4j.data.message.*;
 import dev.langchain4j.model.openai.OpenAiTokenCountEstimator;
-import java.util.ArrayDeque;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
 import java.util.Locale;
 import java.util.Map;
-import java.util.regex.Pattern;
 import java.util.stream.Collectors;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
@@ -273,111 +271,5 @@ public class Messages {
                 .map(Messages::getReprForDisplay)
                 .filter(s -> !s.isBlank())
                 .collect(Collectors.joining("\n\n"));
-    }
-
-    /**
-     * One slice of pre-rendered task markdown after legacy framing has been stripped.
-     * Produced by {@link #parseLegacyFraming(String)}.
-     */
-    public record FramedSegment(ChatMessageType type, String content) {}
-
-    private static final Pattern LEGACY_OPEN_TAG = Pattern.compile("^\\s*<message type=(\\w+)>\\s*$");
-
-    private static final Pattern LEGACY_CLOSE_TAG = Pattern.compile("^\\s*</message>\\s*$");
-
-    private static final Pattern LEGACY_SECTION_LABEL = Pattern.compile("(?m)^(?:Reasoning|Text|Tool calls):\\s*$");
-
-    private static final int MAX_FRAMING_DEPTH = 32;
-
-    /**
-     * Parses task markdown that may contain legacy {@code <message type=X>...</message>} framing
-     * (produced by {@link #format(List)}) into a list of structured segments.
-     *
-     * <p>If no framing is detected the whole markdown is returned as a single CUSTOM segment so that
-     * post-fix data flows through unchanged. This is meant for display-side replay where structured
-     * messages are no longer available on the fragment but the persisted markdown still carries them.
-     *
-     * <p>Uses a line-based stack scanner so it correctly handles empty bodies and arbitrarily nested
-     * framing. Each closing tag emits a segment if its accumulated content is non-empty after cleaning;
-     * nested wrappers contribute additional segments only when they carry their own content outside the
-     * inner blocks. Emission order follows close-tag order (inner blocks emit before their parent).
-     *
-     * <p>If the open-tag depth exceeds {@value #MAX_FRAMING_DEPTH}, the parser falls back to a single
-     * CUSTOM segment containing the original markdown to prevent unbounded allocation on adversarial
-     * input.
-     */
-    public static List<FramedSegment> parseLegacyFraming(String markdown) {
-        if (markdown.isEmpty()) {
-            return List.of();
-        }
-        if (!markdown.contains("<message type=")) {
-            return List.of(new FramedSegment(ChatMessageType.CUSTOM, markdown));
-        }
-        var typeStack = new ArrayDeque<ChatMessageType>();
-        var contentStack = new ArrayDeque<StringBuilder>();
-        var result = new ArrayList<FramedSegment>();
-        for (var line : markdown.split("\n", -1)) {
-            var openMatcher = LEGACY_OPEN_TAG.matcher(line);
-            if (openMatcher.matches()) {
-                if (typeStack.size() >= MAX_FRAMING_DEPTH) {
-                    logger.warn(
-                            "parseLegacyFraming: depth limit {} exceeded; falling back to CUSTOM segment",
-                            MAX_FRAMING_DEPTH);
-                    return List.of(new FramedSegment(ChatMessageType.CUSTOM, markdown));
-                }
-                ChatMessageType type;
-                try {
-                    type = ChatMessageType.valueOf(openMatcher.group(1).toUpperCase(Locale.ROOT));
-                } catch (IllegalArgumentException ex) {
-                    type = ChatMessageType.CUSTOM;
-                }
-                typeStack.push(type);
-                contentStack.push(new StringBuilder());
-                continue;
-            }
-            if (LEGACY_CLOSE_TAG.matcher(line).matches()) {
-                if (typeStack.isEmpty()) {
-                    continue;
-                }
-                var type = typeStack.pop();
-                var cleaned = cleanFramedContent(contentStack.pop().toString());
-                if (!cleaned.isEmpty()) {
-                    result.add(new FramedSegment(type, cleaned));
-                }
-                continue;
-            }
-            if (!contentStack.isEmpty()) {
-                contentStack.peek().append(line).append('\n');
-            }
-        }
-        if (result.isEmpty()) {
-            return List.of(new FramedSegment(ChatMessageType.CUSTOM, markdown));
-        }
-        return result;
-    }
-
-    private static String cleanFramedContent(String content) {
-        if (content.isEmpty()) {
-            return "";
-        }
-        var minIndent = content.lines()
-                .filter(line -> !line.isBlank())
-                .mapToInt(line -> {
-                    int n = 0;
-                    while (n < line.length() && line.charAt(n) == ' ') n++;
-                    return n;
-                })
-                .min()
-                .orElse(0);
-        var deindented = content.lines()
-                .map(line -> line.length() >= minIndent ? line.substring(minIndent) : line)
-                .collect(Collectors.joining("\n"));
-        var withoutLabels = LEGACY_SECTION_LABEL.matcher(deindented).replaceAll("");
-        return withoutLabels.replaceAll("\\n{3,}", "\n\n").strip();
-    }
-
-    /** Strips legacy framing while preserving readable content; used where per-segment structure isn't needed. */
-    public static String stripLegacyFraming(String markdown) {
-        return parseLegacyFraming(markdown).stream().map(FramedSegment::content).collect(Collectors.joining("\n\n"));
     }
 }
